move some stuff to the oss folder as those file create issues later on
