+++ frontend/src/metabase/embedding-sdk/lib/saml-token-storage.ts
@@ -1,4 +1,4 @@
-import type { MetabaseEmbeddingSessionToken } from "embedding-sdk-bundle/types/refresh-token";
+import type { MetabaseEmbeddingSessionToken } from "metabase/embedding-sdk/types/refresh-token";
 
 class TypedStorage<T> {
   constructor(private key: string) {}
